ml-service/requirements.txt
# LeadPulse ML Service Dependencies
# Core
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.3.0

# ML / Data Science
scikit-learn>=1.4.0
xgboost>=2.0.0
prophet>=1.1.5
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0

# NLP / AI
transformers>=4.40.0
torch>=2.2.0
anthropic>=0.25.0
openai>=1.20.0

# RAG
langchain>=0.2.0
langchain-anthropic>=0.1.0

# Utilities
gunicorn>=21.2.0
